+++ front/src/components/contextData/climate/ClimateContextDataMetrics.vue
@@ -2,9 +2,8 @@
 import { computed } from "vue"
 import { type ClimateData } from "@/types/climate"
 import { useClimateZone } from "@/composables/useClimateZone"
-import ContextDataAccordionItem, {
-  type ContextDataFactorGroup
-} from "@/components/contextData/shared/ContextDataAccordionItem.vue"
+import ContextDataAccordionItem from "@/components/contextData/shared/ContextDataAccordionItem.vue"
+import type { ContextDataFactorGroup } from "@/types/contextData"
 
 interface ClimateMetricsProps {
   data: ClimateData
